vismay_requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==1.2.0
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
cuda-bindings @ file:///home/conda/feedstock_root/build_artifacts/cuda-python-split_1737988122893/work/cuda_bindings
cuda-python @ file:///home/conda/feedstock_root/build_artifacts/cuda-python-split_1738524837018/work/cuda_python
cudf @ file:///opt/conda/conda-bld/work/python/cudf
cupy @ file:///home/conda/feedstock_root/build_artifacts/cupy-split_1742850044780/work
cymem==2.0.11
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
executing==2.2.0
faiss-gpu==1.7.2
fastjsonschema==2.21.1
fastrlock @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastrlock_1734873426/work
filelock @ file:///croot/filelock_1700591183607/work
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langchain==0.3.19
langchain-core==0.3.39
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
llvmlite @ file:///croot/llvmlite_1741209858218/work
marisa-trie==1.2.1
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib-inline==0.1.7
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
mistune==3.1.2
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
notebook==7.3.2
notebook_shim==0.2.4
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1739224673889/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx @ file:///home/conda/feedstock_root/build_artifacts/nvtx_1740143162424/work
orjson==3.10.15
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1715897614105/work
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1738010226202/work
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pynvjitlink-cu12 @ file:///opt/conda/conda-bld/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1738317518727/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1733342254348/work/dist
rmm @ file:///opt/conda/conda-bld/work/python/rmm
rpds-py==0.23.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1737133630106/work
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.18.3
zstandard==0.23.0
